% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dawn_return.R
\name{assign_bats}
\alias{assign_bats}
\title{Assign individuals to roosts subject to capacity constraints}
\usage{
assign_bats(n_bats, cell_probability, roost_probability, N_current, N_max)
}
\arguments{
\item{n_bats}{Number of individuals to assign.}

\item{cell_probability}{Probability of selecting each spatial cell.}

\item{roost_probability}{Matrix giving roost probabilities for each
spatial cell.}

\item{N_current}{Current number of individuals occupying each roost.}

\item{N_max}{Maximum capacity of each roost.}
}
\value{
Integer vector containing the number of newly assigned
individuals at each roost.
}
\description{
Assigns individuals to roosts based on cell-specific roost probabilities
while preventing roost capacities from being exceeded.
}
