% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_decay.R
\name{cell_roost_distances}
\alias{cell_roost_distances}
\title{Calculate distances from grid cells to roosts}
\usage{
cell_roost_distances(grid, roost_info)
}
\arguments{
\item{grid}{Spatial grid containing x and y coordinates.}

\item{roost_info}{Data frame containing roost coordinates.}
}
\value{
A matrix of distances between grid cells and roosts.
}
\description{
Calculate distances from grid cells to roosts
}
