% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{create_epidemic_params}
\alias{create_epidemic_params}
\title{Create epidemic parameters}
\usage{
create_epidemic_params(beta = 0.5, gamma = 0.1, model = "SIR", omega = NULL)
}
\arguments{
\item{beta}{Transmission rate.}

\item{gamma}{Recovery rate.}

\item{model}{Epidemic model, either \code{"SIR"} or \code{"SIRS"}.}

\item{omega}{Rate of loss of immunity for the SIRS model.}
}
\value{
A list of epidemic parameters.
}
\description{
Create epidemic parameters
}
