% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{create_movement_params}
\alias{create_movement_params}
\title{Create movement parameters}
\usage{
create_movement_params(
  lambda = 5,
  sigma = 0.01,
  D_base = 5e-04,
  gamma_D = 1,
  alpha = 1,
  epsilon = 0.01,
  times_night = seq(0, 10, 0.1)
)
}
\arguments{
\item{lambda}{Distance-decay parameter.}

\item{sigma}{Initial spatial density spread around roosts.}

\item{D_base}{Baseline diffusion coefficient.}

\item{gamma_D}{Controls the effect of roost isolation on diffusion.}

\item{alpha}{Controls population-based roost attraction.}

\item{epsilon}{Small value preventing zero attraction.}

\item{times_night}{Time sequence for the nighttime diffusion simulation.}
}
\value{
A list of movement parameters.
}
\description{
Create movement parameters
}
