% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_roosts.R
\name{create_roosts}
\alias{create_roosts}
\title{Create a set of roosts}
\usage{
create_roosts(
  num_roosts = 25,
  num_clusters = 5,
  sd = 0.05,
  avg_N_max = 70,
  min_N_max = 10
)
}
\arguments{
\item{num_roosts}{Number of roosts.}

\item{num_clusters}{Number of spatial clusters.}

\item{sd}{Standard deviation of roost locations around cluster centers.}

\item{avg_N_max}{Mean roost carrying capacity.}

\item{min_N_max}{Minimum roost carrying capacity.}
}
\value{
A data frame containing roost locations, carrying capacities,
and population compartments.
}
\description{
Creates spatially clustered roosts with variable carrying capacities.
}
