% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{create_spatial_params}
\alias{create_spatial_params}
\title{Create spatial model parameters}
\usage{
create_spatial_params(
  x_min = 0,
  x_max = 1,
  y_min = 0,
  y_max = 1,
  nx = 100,
  ny = 100
)
}
\arguments{
\item{x_min}{Minimum x coordinate.}

\item{x_max}{Maximum x coordinate.}

\item{y_min}{Minimum y coordinate.}

\item{y_max}{Maximum y coordinate.}

\item{nx}{Number of grid cells in the x direction.}

\item{ny}{Number of grid cells in the y direction.}
}
\value{
A list containing the spatial grid and associated parameters.
}
\description{
Create spatial model parameters
}
