% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{diffusion_SIR}
\alias{diffusion_SIR}
\title{Calculate diffusion derivatives for SIR compartments}
\usage{
diffusion_SIR(t, state, parms)
}
\arguments{
\item{t}{Current simulation time.}

\item{state}{Vector containing S, I, and R spatial densities.}

\item{parms}{List containing nx, ny, D_x, D_y, and grid_2D.}
}
\value{
A list containing the derivatives of S, I, and R.
}
\description{
Calculates spatial diffusion of susceptible, infected, and recovered
individuals using ReacTran.
}
