% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_decay.R
\name{distance_decay}
\alias{distance_decay}
\title{Calculate exponential distance decay}
\usage{
distance_decay(distance, lambda)
}
\arguments{
\item{distance}{Distance values.}

\item{lambda}{Distance-decay parameter.}
}
\value{
Distance-decay weights.
}
\description{
Calculate exponential distance decay
}
