% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_roosts.R
\name{initialize_roost_populations}
\alias{initialize_roost_populations}
\title{Initialize roost populations}
\usage{
initialize_roost_populations(
  roost_info,
  initial_occupancy = 0.25,
  initial_infected = 0.25
)
}
\arguments{
\item{roost_info}{Data frame containing roost capacities.}

\item{initial_occupancy}{Proportion of capacity initially occupied.}

\item{initial_infected}{Proportion of the initial population that
is infected.}
}
\value{
Updated roost information data frame.
}
\description{
Assigns an initial population to each roost and divides individuals
between susceptible and infected compartments.
}
