% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement.R
\name{make_cell_probability}
\alias{make_cell_probability}
\title{Convert a spatial density to cell probabilities}
\usage{
make_cell_probability(cell_density)
}
\arguments{
\item{cell_density}{Matrix of spatial density values.}
}
\value{
Matrix of probabilities with the same dimensions as
\code{cell_density}.
}
\description{
Converts a spatial density surface into probabilities that sum to one.
}
