% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{make_diffusion_faces}
\alias{make_diffusion_faces}
\title{Create diffusion coefficients at grid-cell faces}
\usage{
make_diffusion_faces(D_field)
}
\arguments{
\item{D_field}{Matrix of diffusion coefficients at grid-cell centers.}
}
\value{
A list containing x- and y-direction face coefficients.
}
\description{
Converts a cell-centered diffusion field into x- and y-direction
diffusion coefficients required by ReacTran.
}
