% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{make_diffusion_field}
\alias{make_diffusion_field}
\title{Create a spatially varying diffusion field}
\usage{
make_diffusion_field(cell_roost_distances, D_roost, lambda = 5)
}
\arguments{
\item{cell_roost_distances}{Matrix of distances between grid cells
and roosts.}

\item{D_roost}{Vector of roost-specific diffusion coefficients.}

\item{lambda}{Distance-decay parameter.}
}
\value{
A vector containing the diffusion coefficient for each grid cell.
}
\description{
Calculates the diffusion coefficient at each spatial grid cell by
weighting roost-specific diffusion coefficients according to the
distance-decay relationship between cells and roosts.
}
