% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_decay.R
\name{roost_connectivity}
\alias{roost_connectivity}
\title{Calculate roost connectivity}
\usage{
roost_connectivity(dist_matrix, lambda)
}
\arguments{
\item{dist_matrix}{Matrix of distances between roosts.}

\item{lambda}{Distance-decay parameter.}
}
\value{
A list containing connectivity information.
}
\description{
Calculate roost connectivity
}
