% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_roosts.R
\name{roost_distance_matrix}
\alias{roost_distance_matrix}
\title{Calculate pairwise roost distances}
\usage{
roost_distance_matrix(roost_info)
}
\arguments{
\item{roost_info}{Data frame containing \code{x} and \code{y} roost coordinates.}
}
\value{
A matrix containing pairwise Euclidean distances between roosts.
}
\description{
Calculates Euclidean distances between all pairs of roosts.
}
