% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{run_day}
\alias{run_day}
\title{Run one complete model day}
\usage{
run_day(roost_info, spatial_params, movement_params, epidemic_params)
}
\arguments{
\item{roost_info}{Roost population data frame.}

\item{spatial_params}{Spatial model parameters.}

\item{movement_params}{Movement model parameters.}

\item{epidemic_params}{Epidemic model parameters.}
}
\value{
A list containing roost populations, nighttime densities,
and dawn return probabilities.
}
\description{
Run one complete model day
}
