% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement.R
\name{run_night}
\alias{run_night}
\title{Run nighttime movement}
\usage{
run_night(roost_info, spatial_params, movement_params)
}
\arguments{
\item{roost_info}{Roost population data frame.}

\item{spatial_params}{Spatial model parameters.}

\item{movement_params}{Movement model parameters.}
}
\value{
A list containing the final spatial densities for S, I, and R.
}
\description{
Simulates nighttime movement of susceptible, infected, and recovered
individuals using spatial diffusion.
}
