% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run a multi-day simulation}
\usage{
run_simulation(
  roost_info,
  spatial_params,
  movement_params,
  epidemic_params,
  n_days = 30
)
}
\arguments{
\item{roost_info}{Initial roost populations.}

\item{spatial_params}{Spatial model parameters.}

\item{movement_params}{Movement model parameters.}

\item{epidemic_params}{Epidemic model parameters.}

\item{n_days}{Number of simulation days.}
}
\value{
A list containing model output for each day.
}
\description{
Run a multi-day simulation
}
