% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemiology.R
\name{SIR_step}
\alias{SIR_step}
\title{Perform one SIR epidemiological time step}
\usage{
SIR_step(y, beta, gamma)
}
\arguments{
\item{y}{Numeric vector containing S, I, and R.}

\item{beta}{Transmission rate.}

\item{gamma}{Recovery rate.}
}
\value{
Numeric vector containing updated S, I, and R.
}
\description{
Updates susceptible, infected, and recovered individuals in a roost using binomial transitions for infection and recovery.
}
