% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemiology.R
\name{SIRS_step}
\alias{SIRS_step}
\title{Perform one SIRS epidemiological time step}
\usage{
SIRS_step(y, beta, gamma, omega)
}
\arguments{
\item{y}{Numeric vector containing S, I, and R.}

\item{beta}{Transmission rate.}

\item{gamma}{Recovery rate.}

\item{omega}{Rate of loss of immunity.}
}
\value{
Numeric vector containing updated S, I, and R.
}
\description{
Updates susceptible, infected, and recovered individuals in a roost,
including loss of immunity from the recovered compartment.
}
